+++ Firestore/Example/Tests/Integration/FSTTransactionTests.mm
@@ -25,6 +25,8 @@
 using firebase::firestore::util::TimerId;
 
 @interface FSTTransactionTests : FSTIntegrationTestCase
+- (void)runFailedPreconditionTransactionWithOptions:(FIRTransactionOptions *_Nullable)options
+                                  expectNumAttempts:(int)expectedNumAttempts;
 @end
 
 /**
@@ -793,6 +795,64 @@ TransactionStage get = ^(FIRTransaction *transaction, FIRDocumentReference *doc)
         [expectation fulfill];
       }];
   [self awaitExpectations];
+}
+
+- (void)runFailedPreconditionTransactionWithOptions:(FIRTransactionOptions *_Nullable)options
+                                  expectNumAttempts:(int)expectedNumAttempts {
+  // Note: The logic below to force retries is heavily based on
+  // testRetriesWhenDocumentThatWasReadWithoutBeingWrittenChanges.
+
+  FIRFirestore *firestore = [self firestore];
+  FIRDocumentReference *doc = [[firestore collectionWithPath:@"counters"] documentWithAutoID];
+  auto attemptCount = std::make_shared<std::atomic_int>(0);
+  attemptCount->store(0);
+
+  [self writeDocumentRef:doc data:@{@"count" : @"initial value"}];
+
+  // Skip backoff delays.
+  [firestore workerQueue]->SkipDelaysForTimerId(TimerId::RetryTransaction);
+
+  XCTestExpectation *expectation = [self expectationWithDescription:@"transaction"];
+  [firestore runTransactionWithOptions:options
+      block:^id _Nullable(FIRTransaction *transaction, NSError **error) {
+        ++(*attemptCount);
+
+        [transaction getDocument:doc error:error];
+        XCTAssertNil(*error);
+
+        // Do a write outside of the transaction. This will force the transaction to be retried.
+        dispatch_semaphore_t writeSemaphore = dispatch_semaphore_create(0);
+        [doc setData:@{
+          @"count" : @(attemptCount->load())
+        }
+            completion:^(NSError *) {
+              dispatch_semaphore_signal(writeSemaphore);
+            }];
+        dispatch_semaphore_wait(writeSemaphore, DISPATCH_TIME_FOREVER);
+
+        // Now try to update the doc from within the transaction.
+        // This will fail since the document was modified outside of the transaction.
+        [transaction setData:@{@"count" : @"this write should fail"} forDocument:doc];
+        return nil;
+      }
+      completion:^(id, NSError *_Nullable error) {
+        [self assertError:error
+                  message:@"the transaction should fail due to retries exhausted"
+                     code:FIRFirestoreErrorCodeFailedPrecondition];
+        XCTAssertEqual(attemptCount->load(), expectedNumAttempts);
+        [expectation fulfill];
+      }];
+  [self awaitExpectations];
+}
+
+- (void)testTransactionOptionsNil {
+  [self runFailedPreconditionTransactionWithOptions:nil expectNumAttempts:5];
+}
+
+- (void)testTransactionOptionsMaxAttempts {
+  FIRTransactionOptions *options = [[FIRTransactionOptions alloc] init];
+  options.maxAttempts = 7;
+  [self runFailedPreconditionTransactionWithOptions:options expectNumAttempts:7];
 }
 
 @end