+++ Firestore/core/test/unit/local/counting_query_engine.h
@@ -58,7 +58,7 @@ class CountingQueryEngine : public QueryEngine {
 
   /**
    * Returns the number of documents returned by the RemoteDocumentCache's
-   * `GetMatching()` API (since the last call to `ResetCounts()`)
+   * `GetAll()` API (since the last call to `ResetCounts()`)
    */
   size_t documents_read_by_query() const {
     return documents_read_by_query_;
@@ -73,12 +73,10 @@ class CountingQueryEngine : public QueryEngine {
   }
 
   /**
-   * Returns the number of mutations returned by the MutationQueue's
-   * `getAllMutationBatchesAffectingQuery()` API (since the last call to
-   * `ResetCounts()`)
+   * Returns the number of overlays returned by the DocumentOverlayCache.
    */
-  size_t mutations_read_by_query() const {
-    return mutations_read_by_query_;
+  size_t overlays_read_by_collection() const {
+    return overlays_read_by_collection_;
   }
 
   /**
@@ -87,8 +85,8 @@ class CountingQueryEngine : public QueryEngine {
    * `AllMutationBatchesAffectingDocumentKeys()` APIs (since the last call to
    * `ResetCounts()`)
    */
-  size_t mutations_read_by_key() const {
-    return mutations_read_by_key_;
+  size_t overlays_read_by_key() const {
+    return overlays_read_by_key_;
   }
 
  private:
@@ -179,9 +177,8 @@ class WrappedRemoteDocumentCache : public RemoteDocumentCache {
 
   model::MutableDocumentMap GetAll(const model::DocumentKeySet& keys) override;
 
-  model::MutableDocumentMap GetMatching(
-      const core::Query& query,
-      const model::SnapshotVersion& since_read_time) override;
+  model::MutableDocumentMap GetAll(const model::ResourcePath& path,
+                                   const model::IndexOffset& offset) override;
 
   void SetIndexManager(IndexManager* manager) override {
     index_manager_ = NOT_NULL(manager);
@@ -205,16 +202,16 @@ class WrappedDocumentOverlayCache final : public DocumentOverlayCache {
       const model::DocumentKey& key) const override;
 
   void SaveOverlays(int largest_batch_id,
-                    const MutationByDocumentKeyMap& overlays) override;
+                    const model::MutationByDocumentKeyMap& overlays) override;
 
   void RemoveOverlaysForBatchId(int batch_id) override;
 
-  OverlayByDocumentKeyMap GetOverlays(const model::ResourcePath& collection,
-                                      int since_batch_id) const override;
+  model::OverlayByDocumentKeyMap GetOverlays(
+      const model::ResourcePath& collection, int since_batch_id) const override;
 
-  OverlayByDocumentKeyMap GetOverlays(absl::string_view collection_group,
-                                      int since_batch_id,
-                                      std::size_t count) const override;
+  model::OverlayByDocumentKeyMap GetOverlays(absl::string_view collection_group,
+                                             int since_batch_id,
+                                             std::size_t count) const override;
 
  private:
   int GetOverlayCount() const override;